Handle first release of new packages
This may or may not have been the case before, but now packagecloud
returns a 404 status code when asking for the versions of a package
which does not yet exist. Our code expected an empty array.

With a minor tweak, we can check for the 404 status code and interpose
an empty array as the response body. Basically we ask that curl print
the http code on a final line after the full response body. The rest is
just shell scripting. This restores normal processing during the first
build of a new package and maintains the old behavior for other status
codes.