Update indent config for uncrustify 0.65 compat.
It applies the struct new line rule slightly differently than before,
and changing it to 1 from 3 results in no changes in our existing base
whether using uncrustify 0.64 or 0.65.

Additionally, 0.65 seems to want to change indentation on some usages
of 'break' outside of bracketed regions, so I'm adding a line that will
force breaks to be put within brackets, which avoids the problem and
makes the codebase more consistent at the same time.