Fix rpm jq version transform for PG >= 10
We had assumed RPMs used a number like '95' or '96' in their package
names, so would add a decimal point between the first and second digit.

Now that 10, 11, 12 exist, this would be wrong, so adding a small regex
can fix it (at least until we actual have versions in the 90s).