add pgbouncer support (#81)
It is non trivial to setup pgbouncer between the coordinators and the workers. Due to the complexities we experiment little with new features having pgbouncer setup in between.

This patch makes it trivial to have pgbouncer setup. During the creation of the cluster you pass the `--with-pgbouncer` flag. This will write the `pgbouncer` configuration files and starts pgbouncer in daemon mode. It does install the records in `pg_dist_poolinfo`, meaning this can only be used with the Citus enterprise repository as of now.