Feature: citus_dev pipenv shebang (#326)
This adds a second script next to `citus_dev`. When a user doesn't want/can install the requirements globally they can opt to install `pipenv` + `pipenv-shebang` and link `citus_dev` on their path to `citus_dev-pipenv` to run the script in its isolated environment.

If the user opts to use `citus_dev-pipenv` it is their responsibility to `pipenv install` or `pipenv sync` in the source directory to make sure the environment is initialized correctly.